Update Visualizing all Phylogenies.R
Okay so when I try to push from Rstudios directly, it gets pushed to test2, which makes sense because all these scripts are in the test2 project. HOWEVER, when I push it from github desktop, it correctly (desirably?) pushes the updates to the Honours-Thesis repo. So....To be honest I think it might be an issue with the weird way you've set up your github, but I think in the future?? both ways should work the same?